Update
Rename revealExpiredBlock to deadlineBlock. When game is first created, deadlineBlock will hold the deadline that creator has to wait in order to be able to cancel his game. When the first player reveal his move, deadlineBlock will hold deadline for the second player to reveal his move. or the first player will win.